feat(loading): enhance splash screen visibility logic
Added a console log to help trace the state changes of the splash screen visibility. This will assist in debugging and ensuring the correct application of the 'splash-screen-hidden' class based on the loading state.

chore(nephos-utils): bump version to 1.2.1-rc.2

Updated the package version to ensure proper versioning and dependency management in future releases. This change will help in tracking modifications and maintaining the package more efficiently.