feat(utils): add checkFor method in loading service
Added a new method `checkFor` to the loading service to check the presence of specific keys. This enhancement aims to improve the flexibility and usability of the loading service by allowing finer control over the loading states.

build(utils): update package version to 1.3.0-rc.1

Updated the package version to 1.3.0-rc.1 to mark the introduction of new features and improvements in the utility package. This version update signifies the ongoing development and enhancement of the utilities.